PR-B-8: extract DeclarationBinder (per-declaration-kind binding + interface-implementation queue)
Extract every per-declaration-kind binder from Binder.cs into a new
internal sealed class DeclarationBinder, owning ~3.7k LoC of
declaration-binding logic plus the pending struct->interface
implementation queue.

Methods moved (grouped by declaration kind):

  Type aliases / delegates / enums
    - BindTypeAliasDeclaration
    - BindDelegateDeclaration
    - BindEnumDeclaration

  Structs (the largest body driver)
    - BindStructDeclaration
    - BindStructDeclarationBody
    - field / property / event declaration helpers
    - BindConstructorDeclarations
    - BindBaseConstructorInitializer
    - VerifyInterfaceImplementations (the drain pass)
    - pendingInterfaceImplementationChecks (field)

  Interfaces
    - DeclareInterfaceSymbol
    - BindInterfaceMembers / BindInterfaceMembersCore

  Functions
    - BindFunctionDeclaration
    - BindTypeParameterList
    - validation helpers (CheckVariancePosition, SignaturesMatch,
      FindRefKindMismatchIndex, GetCallableParameters,
      GetBaseClauseTypeDisplayName, HasUnscopedRefAnnotation)

  Variables (symbol-construction overloads)
    - BindVariableDeclaration(SyntaxToken, bool, TypeSymbol)
    - BindVariableDeclaration(SyntaxToken, bool, TypeSymbol, Accessibility)

  Attributes
    - BindAttributes / BindAttribute

DeclarationBinder receives 15 callbacks from the root Binder
(constructor-injected, never back-references Binder): BindExpression,
BindTypeClause, BindReturnTypeClause, BindTypeOfExpression,
BindArrayCreationExpression, ResolveAccessibility, LookupType,
GetEffectiveArgumentClrType, IsAsyncIteratorReturnType,
IsAsyncSequenceReturnType, IsPrimitiveTypeName, RefKindToString,
GetCurrentFunction, plus a back-reference to ConversionClassifier.

What stays on Binder (and why):

  - BindGlobalScope / BindProgram / BindImport: top-level entry points.
  - TryInferExpressionType: public API used by HoverComputer.
  - The seven *AllowedTargets static hashsets: per plan directive,
    constants stay on root (widened from private to internal static
    readonly so DeclarationBinder can reference them as Binder.X).
  - AttachDocumentation: has non-declaration call sites
    (package/import). Widened to internal static.
  - FormatOverloadSignature: also used by expression-side overload
    resolution diagnostics. Widened to internal static.
  - HasUnscopedRefAnnotation: kept on Binder (with internal static
    forwarder from DeclarationBinder.HasUnscopedRefAnnotation) because
    Binder constructor uses it when seeding receiver-parameter scope.
  - The Bind*Expression family: scoped to PR-B-9.

pendingInterfaceImplementationChecks moved with its drain pass
(VerifyInterfaceImplementations). The drain call site in BindGlobalScope
now invokes declarations.VerifyInterfaceImplementations(structSymbol,
syntax).

Verified:
  - dotnet build GSharp.sln: 0 warnings, 0 errors.
  - dotnet test test/Core.Tests: 2120 passed, 1 skipped, 0 failed.
  - RefactoringBaselineTests: 92/92 passed, zero hash diffs.
  - dotnet test test/Interpreter.Tests / Sdk.Tests / LanguageServer.Tests:
    all green.
  - Compiler.Tests: env-bound failures only (Too many open files /
    OOM on local macOS); the same failures reproduce on origin/main.
    Targeted decl/interface/struct/ctor filters: 9/9 passed.

Public/internal API surface unchanged.

LoC delta:
  - src/Core/CodeAnalysis/Binding/Binder.cs: 12,123 -> 8,433 (-3,690)
  - src/Core/CodeAnalysis/Binding/DeclarationBinder.cs: new, 3,844 LoC

Next: PR-B-9 ExpressionBinder (last + biggest Phase-1 PR).
